backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
google-generativeai==0.3.2
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
